notification mailer: Include other clients in the `To` email field.
To allow clients to choose to which other clients they would like to
reply, include other clients in the `To` email field.

Agents are not included in the `To` email field to prevent their
personal email addresses from becoming public.

This change is discussed in https://github.com/ivaldi/brimir/issues/259.